+++ src/Numeric/Units/Dimensional/SIUnits.hs
@@ -55,17 +55,21 @@ module Numeric.Units.Dimensional.SIUnits
   -- $multiples
   deka, deca, hecto, kilo, mega, giga, tera, peta, exa, zetta, yotta,
   -- $submultiples
-  deci, centi, milli, micro, nano, pico, femto, atto, zepto, yocto
+  deci, centi, milli, micro, nano, pico, femto, atto, zepto, yocto,
+  -- $reified-prefixes
+  Prefix, applyPrefix, siPrefixes
 )
 where
 
+import Data.Ratio
 import Numeric.Units.Dimensional
 import Numeric.Units.Dimensional.Quantities
-import Numeric.Units.Dimensional.UnitNames (PrefixName, applyPrefix, nMeter, nGram, nSecond, nAmpere, nKelvin, nMole, nCandela)
+import Numeric.Units.Dimensional.UnitNames (Prefix, siPrefixes)
 import qualified Numeric.Units.Dimensional.UnitNames as N
 import Numeric.Units.Dimensional.UnitNames.Internal (ucum, ucumMetric)
+import qualified Numeric.Units.Dimensional.UnitNames.Internal as I
 import Numeric.NumType.DK.Integers ( pos3 )
-import Prelude ( ($), Num, Fractional, Floating, Integer, Rational)
+import Prelude ( Eq(..), ($), Num, Fractional, Floating, otherwise, error)
 import qualified Prelude
 
 {- $multiples
@@ -80,42 +84,56 @@ section 6.2.6 "Unacceptability of stand-alone prefixes".
 We define all SI prefixes from Table 5. Multiples first.
 -}
 
-applyMultiple :: (Num a) => PrefixName -> Integer -> Unit 'Metric d a -> Unit 'NonMetric d a
-applyMultiple p x u = mkUnitZ (applyPrefix p (name u)) x u
+applyMultiple :: (Num a) => Prefix -> Unit 'Metric d a -> Unit 'NonMetric d a
+applyMultiple p u | denominator x == 1 = mkUnitZ n' (numerator x) u
+                  | otherwise = error "Attempt to apply a submultiple prefix as a multiple."
+  where
+    n' = N.applyPrefix p (name u)
+    x = N.scaleFactor p
 
 deka, deca, hecto, kilo, mega, giga, tera, peta, exa, zetta, yotta
   :: Num a => Unit 'Metric d a -> Unit 'NonMetric d a
-deka  = applyMultiple N.deka 10 -- International English.
+deka  = applyMultiple I.deka -- International English.
 deca  = deka      -- American English.
-hecto = applyMultiple N.hecto 100
-kilo  = applyMultiple N.kilo 1e3
-mega  = applyMultiple N.mega 1e6
-giga  = applyMultiple N.giga 1e9
-tera  = applyMultiple N.tera 1e12
-peta  = applyMultiple N.peta 1e15
-exa   = applyMultiple N.exa 1e18
-zetta = applyMultiple N.zetta 1e21
-yotta = applyMultiple N.yotta 1e24
+hecto = applyMultiple I.hecto
+kilo  = applyMultiple I.kilo
+mega  = applyMultiple I.mega
+giga  = applyMultiple I.giga
+tera  = applyMultiple I.tera
+peta  = applyMultiple I.peta
+exa   = applyMultiple I.exa
+zetta = applyMultiple I.zetta
+yotta = applyMultiple I.yotta
 
 {- $submultiples
 Then the submultiples.
 -}
 
-applySubmultiple :: (Fractional a) => PrefixName -> Rational -> Unit 'Metric d a -> Unit 'NonMetric d a
-applySubmultiple p x u = mkUnitQ (applyPrefix p (name u)) x u
+applyPrefix :: (Fractional a) => Prefix -> Unit 'Metric d a -> Unit 'NonMetric d a
+applyPrefix p u = mkUnitQ n' x u
+  where
+    n' = N.applyPrefix p (name u)
+    x = N.scaleFactor p
 
 deci, centi, milli, micro, nano, pico, femto, atto, zepto, yocto
   :: Fractional a => Unit 'Metric d a -> Unit 'NonMetric d a
-deci  = applySubmultiple N.deci 0.1
-centi = applySubmultiple N.centi 0.01
-milli = applySubmultiple N.milli 1e-3
-micro = applySubmultiple N.micro 1e-6
-nano  = applySubmultiple N.nano 1e-9
-pico  = applySubmultiple N.pico 1e-12
-femto = applySubmultiple N.femto 1e-15
-atto  = applySubmultiple N.atto 1e-18
-zepto = applySubmultiple N.zepto 1e-21
-yocto = applySubmultiple N.yocto 1e-24
+deci  = applyPrefix I.deci
+centi = applyPrefix I.centi
+milli = applyPrefix I.milli
+micro = applyPrefix I.micro
+nano  = applyPrefix I.nano
+pico  = applyPrefix I.pico
+femto = applyPrefix I.femto
+atto  = applyPrefix I.atto
+zepto = applyPrefix I.zepto
+yocto = applyPrefix I.yocto
+
+{- $reified-prefixes 
+
+We supply an explicit representation of an SI prefix, along with a function to apply one and a 
+list of all prefixes defined by the SI.
+
+-}
 
 {- $base-units
 These are the base units from section 4.1. To avoid a
@@ -128,7 +146,7 @@ We define the SI base units in the order of table 1.
 -}
 
 metre, meter :: Num a => Unit 'Metric DLength a
-metre = mkUnitZ nMeter 1 siUnit -- International English.
+metre = mkUnitZ I.nMeter 1 siUnit -- International English.
 meter = metre         -- American English.
 
 {-
@@ -140,17 +158,17 @@ The drawback is that we are forced to use 'Fractional'.
 -}
 
 gram    :: Fractional a => Unit 'Metric DMass a
-gram    = mkUnitQ nGram 1e-3 siUnit
+gram    = mkUnitQ I.nGram 1e-3 siUnit
 second  :: Num a => Unit 'Metric DTime a
-second  = mkUnitZ nSecond 1 siUnit
+second  = mkUnitZ I.nSecond 1 siUnit
 ampere  :: Num a => Unit 'Metric DElectricCurrent a
-ampere  = mkUnitZ nAmpere 1 siUnit
+ampere  = mkUnitZ I.nAmpere 1 siUnit
 kelvin  :: Num a => Unit 'Metric DThermodynamicTemperature a
-kelvin  = mkUnitZ nKelvin 1 siUnit
+kelvin  = mkUnitZ I.nKelvin 1 siUnit
 mole    :: Num a => Unit 'Metric DAmountOfSubstance a
-mole    = mkUnitZ nMole 1 siUnit
+mole    = mkUnitZ I.nMole 1 siUnit
 candela :: Num a => Unit 'Metric DLuminousIntensity a
-candela = mkUnitZ nCandela 1 siUnit
+candela = mkUnitZ I.nCandela 1 siUnit
 
 {- $derived-units
 From Table 3, SI derived units with special names and symbols, including the