+++ CHANGELOG.md
@@ -6,9 +6,12 @@ vNext
   `HasDynamicDimension` typeclass.
 * Breaking: Added operators for `AnyUnit` to the Numeric.Units.Dimensional.Dynamic
   module which may cause name collisions.
+* Breaking: Removed exports of `nMeter`, `nSecond`, `kilo`, etc from Numeric.Units.Dimensional.UnitNames.
+  Access these instead by inspecting the relevant units or prefixes.
 * Added `Data`, `Generic`, `Typeable` and `NFData` instances for many ancillary types.
 * Added `unQuantity` to the Coercion module to ease unwrapping without
   introducing ambiguous type variables.
+* Created explicit representation of metric `Prefix`es.
 * Added a multiplicative `Monoid` instance for `AnyQuantity` and for `AnyUnit`.
 * Added the `DynQuantity` type to represent possibly valid quantities of statically
   unknown dimension, suitable for performing arithmetic with such quantities.